Move the stance report into the library and share the SB3 checkpoint loader
Phase K (cleanup), report-script batch — entries DU6, ST7 and the
foot_sensor_report leftover of DU3 from
docs/reviews/RL_PIPELINE_GAP_REVIEW_2026_08.md §6.2/§6.4. Hand-runs of every
touched script are byte-identical before and after.

ST7: reporting/stage_artifacts.py imported report builders from
scripts/stance_gate_report.py at eight call sites, inverting the package
boundary (pyproject excludes */scripts/*). Everything but main() and the
argparse block moves verbatim into reporting/stance_report.py; the script
keeps the CLI and re-imports the public names, so the command, its docs
and every test keep working, and importing the library no longer loads
the scripts package (pinned by a meta-path test).

DU6: the PPO.load + VecNormalize sidecar block was copy-pasted across the
stance gate report and three diagnostic scripts, with the certification
tool carrying the plant-aware, diagnosable-error variant.
policy_loading.load_sb3_checkpoint is the one implementation; the four
callers pass their genuine differences (sidecar guessing, fail-closed,
plant validation) as parameters. The one observable change is that the
three scripts now report a corrupt sidecar with the certification tool's
diagnosable message instead of a raw unpickling traceback (still exit 1).

DU3 leftover: foot_sensor_report builds its envs through config.build_env,
which gains optional env overrides for its reset_noise_scale=0.0.
