+++ environments/shared/scripts/action_bound_report.py
@@ -4,10 +4,11 @@
 The action space is ``Box(-1, 1)``, but SB3's Gaussian policy is unbounded, so
 the sampled action routinely lands outside it.  Both training paths clip before
 stepping the environment -- SB3 at ``on_policy_algorithm.py:214-218`` and
-``policies.py:379`` (inside ``predict``), the JAX trainer at
-``jax_trainer.py:365`` -- so the plant and the reward both see an in-bound
-action.  What the clip hides is how hard the policy is pushing against the
-bound, and that is what this script measures.
+``policies.py:379`` (inside ``predict``), the JAX trainer in
+``collect_rollout`` inside ``_build_jit_fns``
+(``environments/shared/jax_train_fn.py``) -- so the plant and the reward both
+see an in-bound action.  What the clip hides is how hard the policy is pushing
+against the bound, and that is what this script measures.
 
 It matters because PPO stores the *raw* action and its ``log_prob`` while the
 environment responds to the *clipped* one.  Once a large fraction of components
@@ -73,38 +74,11 @@ if _repo_root not in sys.path:
 
 from environments.shared.config import SPECIES_NAMES, build_env
 from environments.shared.constants import PUBLICATION_SEED_START
+from environments.shared.policy_loading import UNNORMALIZED_BANNER, PolicyLoadError, load_sb3_checkpoint
 
 # Matches DiagnosticsCallback.action_saturation_threshold, so the figures here
 # are directly comparable to diagnostics/raw_action_saturation.
 SATURATION_THRESHOLD = 0.99
-UNNORMALIZED_BANNER = "UNNORMALIZED EVAL — results are not comparable to training-time metrics"
-
-
-def resolve_vecnorm_path(model_path: str, vecnorm_arg: str | None, allow_unnormalized: bool) -> str | None:
-    """The VecNormalize sidecar to evaluate with, or ``None`` for a deliberately unnormalised run.
-
-    An explicit ``--vecnorm`` wins.  Otherwise the trainer's own resolver
-    probes both sidecar conventions -- the ``<stem>_vecnorm.pkl`` guess this
-    script used to make can never match SB3's periodic
-    ``<prefix>_vecnormalize_<steps>_steps.pkl``, so every periodic checkpoint
-    was silently scored on raw observations.  No sidecar is fatal unless
-    ``--allow-unnormalized``: a policy evaluated unnormalised is a different
-    policy, and its saturation figures describe nothing that trained.
-    """
-    if vecnorm_arg is not None:
-        return vecnorm_arg
-    from environments.shared.train_base import _resolve_vecnorm_sidecar
-
-    candidate = _resolve_vecnorm_sidecar(model_path)
-    if Path(candidate).exists():
-        return candidate
-    if allow_unnormalized:
-        return None
-    raise SystemExit(
-        f"no VecNormalize sidecar found for {model_path} (probed {candidate}). A policy evaluated on "
-        "unnormalised observations is a different policy; pass --vecnorm, or --allow-unnormalized "
-        "to proceed deliberately."
-    )
 
 
 def actuator_names(env, n_actuators: int) -> list[str]:
@@ -138,15 +112,16 @@ def main(argv: list[str]) -> None:
     model = None
     normalizer = None
     if args.model:
-        from stable_baselines3 import PPO
-        from stable_baselines3.common.vec_env import DummyVecEnv, VecNormalize
-
-        model = PPO.load(args.model, device="cpu")
-        vecnorm_path = resolve_vecnorm_path(args.model, args.vecnorm, args.allow_unnormalized)
-        if vecnorm_path:
-            normalizer = VecNormalize.load(vecnorm_path, DummyVecEnv([lambda: build_env(args.species, args.stage)]))
-            normalizer.training = False
-            normalizer.norm_reward = False
+        try:
+            model, normalizer, _ = load_sb3_checkpoint(
+                args.model,
+                args.vecnorm,
+                lambda: build_env(args.species, args.stage),
+                allow_unnormalized=args.allow_unnormalized,
+            )
+        except PolicyLoadError as exc:
+            # The CLI boundary is where a loading failure becomes an exit status.
+            raise SystemExit(str(exc)) from exc
 
     rng = np.random.default_rng(args.seed)
     returns: list[float] = []
